src:4.25 Specular light 4.27 Point light

+++ LearnOGL/Window.h
@@ -12,8 +12,8 @@ public:
 
 	int initialize();
 
-	GLfloat getBufferWidth() { return bufferWidth; }
-	GLfloat getBufferHeight() { return bufferHeight; }
+	GLint getBufferWidth() { return bufferWidth; }
+	GLint getBufferHeight() { return bufferHeight; }
 
 	bool getShouldClose() {
 		return glfwWindowShouldClose(mainWindow);
